docker/model_training/requirements.txt
mlflow==2.11.0
scikit-learn==1.4.0
numpy==1.26.4
python-dotenv
